refactor(nodes): reuse `NodeStatusCheckbox` component
