+++ src/components/nodes/adm/index.ts
@@ -1,1 +1,1 @@
-export { default as NodesTable } from './NodesTable/NodesTable.vue'
+export { default as AdmNodesTable } from './AdmNodesTable.vue'
